airframe-sql: Support in-expression sub-query in traverse and transform (#2698)
This will make possible to traverse/transform `LogicalPlan` in
`Expression`.

Another idea is moving all traverse and transform methods (including
`mapChildren`) to TreeNode with a signature change `LogicalPlan` ->
`TreeNode[_]`. This will change the ability of those methods that can
traverse/transform any nodes in the tree, not only `LogicalPlan`.
Meaning, no need to have methods for `LogicalPlan` and `Expression`
independently.

I'm not sure which is better. 🤔