airframe-sql: Support EXTRACT function  (#2818)
Currently, queries containing `EXTRACT` cannot properly be parsed as
`SQLInterpreter` does not override the `visitExtract` method. The
default one ends up producing a `null` value when a function call is
made inside the `EXTRACT` statement. An exception is later thrown by the
parser when it encounters that `null` value.

This PR adds support for the `EXTRACT` functions. It also adds support
for additional intervals that need to be supported and it fixes an issue
when generating grouping keys; function call would not be printed
correctly.