airframe-sql: Fix equi self-join (#2819)
Certain queries containing equi self-join cannot be parsed:
```
select A.name
from A
inner join (select name from A) t2 
on A.name = t2.name
```
The type resolver will end up throwing an exception, complaining about
ambiguous column references. This is because the source code tries to
match columns based on their source table. I guess it's to handle joins
with unqualified attributes.

It took me a while to understand what the code was doing. I think that
the matcher should not look at the source table name if a qualifier is
present.