airframe-sql: Support SUBSTRING (#2830)
SUBSTRING function calls are not properly supported. For example, the
following query cannot be parsed at the moment:
```
SELECT SUBSTRING(a FROM 1 FOR 5) FROM b
```

With this PR, `SUBSTRING` function calls are handled as regular function
calls. This is safe as `SUBSTRING(a FROM 1 FOR 5)` is equivalent to
`SUBSTRING(a, 1, 5)`.

---------

Co-authored-by: Taro L. Saito <leo@xerial.org>