fix(sql): Improve quotation handling (#2964)
Quotation marks are not really well handled at the moment. Sometimes,
they're stripped, sometimes they're not. Rather than introducing massive
changes to preserve the fact that an identifier was originally quotes, I
propose the strip them whenever they end up being stored in a `String`.
On the generator side, we can than automatically add the quotes back if
we detect that they are needed.