fix: Various fixes to make http-recorder usable again.  (#2983)
As it stands, http-recorder only works with GET requests without a body.
This situation is caused by two issues:
- The hash code of the body is not computed correctly. `Arrays.hashcode`
should be used to compute an hash code over the content of the array as
the `hashcode` method on the array itself only relies on the address of
the object.
- `RedirectToRxEndpoint` only exposes a `GET` endpoint. No other verbs
can pass-through. I don't know if there's a better way to fix it, but
the quick fix I'm proposing works.