Support sslmode parameter in PostgreSQL (#3001)
## Overview

- Support `sslmode` parameter in PostgreSQL


## Background

- For relax verification when connecting to a DB using a CNAME record.
- When using a RDS host, a RDS is set up with a SSL certificate managed
by AWS. But, a hostname is generated with randomly. It's not convenient.
So, some of the users use DNS to set up an arbitrary name and use it to
connect to the DB.
- At that case, the default sslmode gets cause of the error since a
proper SSL certification isn't used.
- For the `psql` command, it uses `prefer` when connecting to a DB with
ssl. And, it works as expected.

<!--
- `pgJDBC` and `psql`, which is a default CLI commnd, use different
sslmode by default.
`psql` use `prefer` but `pgJDBC` use `verify-full` when connecting to a
DB with ssl.
-->