internal: Use Scala 3 by default for sbt and IntelliJ (#3137)
As the primary development has moved to Scala 3, adding Scala 3 source
code (scala-3) to the IntelliJ project is better for daily development.
We still need to support Scala 2 for a while, but we don't have so many
Scala 2 specific source code (scala-2 folder). This change is to let
IntelliJ recognize scala-3 source code folder to enable syntax
highlight, etc.